+++ lib/helpers/dependencies_injection.dart
@@ -3,7 +3,8 @@
 /// dependencies registration and injection
 import 'package:get_it/get_it.dart';
 import 'package:xiaokeai/helpers/get_platform_service.dart';
-import 'package:xiaokeai/services/configs/permission_service.dart';
+import 'package:xiaokeai/services/auth/local/local_auth_service.dart';
+import 'package:xiaokeai/services/configs/utils/permission_service.dart';
 import 'package:xiaokeai/services/package/package_info_provider.dart';
 import 'package:xiaokeai/services/package/package_info_service.dart';
 import 'package:xiaokeai/services/pref/shared_pref_service.dart';
@@ -21,8 +22,9 @@ Future<void> setupDependencies() async {
   getIt.registerLazySingleton(
       () => PackageInfoProvider(getIt<PackageInfoService>()));
   await setupSharedPref();
-  setupPermission();
+  getIt.registerSingleton<PermissionsService>(PermissionsService());
   getIt.registerSingleton<PlatformDetectionService>(PlatformDetectionService());
+  getIt.registerLazySingleton<LocalAuthService>(() => LocalAuthService());
   // getIt.registerLazySingleton(() => SpeechService());
 }
 
@@ -30,8 +32,4 @@ Future<void> setupSharedPref() async {
   final sharedPrefsService = SharedPreferencesService();
   await sharedPrefsService.init();
   getIt.registerSingleton<SharedPreferencesService>(sharedPrefsService);
-}
-
-void setupPermission() {
-  getIt.registerSingleton<PermissionsService>(PermissionsService());
 }