+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Initial Development: Setup Firebase for the app, added logger system, added the permission system + notification system & setting system (language), and other improvements on September 27, 2024 -> **0.0.2** by Luke
 - Initial Development: Added launch screen, and other improvements on September 28, 2024 -> **0.0.3** by Luke
 - Initial Development: Added Theme and l10n in actions, and other improvements on September 28, 2024 -> **0.0.4** by Luke
+- Initial Development: Setup the underlying global error exceptions, added local auth service file, and other improvements on September 28, 2024 -> **0.0.5** by Luke